carry the policy weight in 64 bits without growing the node
An entry's value is written under the node's monitor but its update
task is queued after the monitor is released, so a second writer can
queue first and the replay applies the deltas out of order. The node's
policy weight then walks past the true value and back, which is
tolerated. Once the intermediate sum leaves the int range, though, a
transfer copies the truncated value into the long region totals and no
later task repairs that copy: the node settles at its real weight while
the region keeps a permanent 2^32 residue. The window then reads as
over its maximum, is drained every maintenance cycle, and every later
arrival goes straight to probation.

Guarding the transfer on a negative weight does not close it, since
half the wraps land positive and read as an ordinary weight. Widening
the field would grow the sixteen strong-value weighted node classes by
eight bytes, so the low half stays in its own field and the high half
moves into the queue word, renamed metadata for the bag of bits it now
is. JOL reports all 147 node classes unchanged.

+++ caffeine/src/main/java/com/github/benmanes/caffeine/cache/BoundedLocalCache.java
@@ -712,7 +712,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
       }
 
       Node<K, V> next = node.getNextInAccessOrder();
-      int weight = node.getPolicyWeight();
+      long weight = node.getPolicyWeight();
       if (weight != 0) {
         transfer(node, weight, WINDOW, PROBATION);
         if (first == null) {
@@ -1175,7 +1175,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
         break;
       }
 
-      int weight = candidate.getPolicyWeight();
+      long weight = candidate.getPolicyWeight();
       if (quota < weight) {
         break;
       }
@@ -1207,7 +1207,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
         break;
       }
 
-      int weight = candidate.getPolicyWeight();
+      long weight = candidate.getPolicyWeight();
       if (quota < weight) {
         break;
       }
@@ -1245,7 +1245,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
    * region that tracks its own.
    */
   @GuardedBy("evictionLock")
-  void transfer(Node<K, V> node, int weight, int from, int to) {
+  void transfer(Node<K, V> node, long weight, int from, int to) {
     if (from == WINDOW) {
       setWindowWeightedSize(windowWeightedSize() - weight);
       accessOrderWindowDeque().remove(node);
@@ -2110,7 +2110,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
         if (access.recordsMiss()) {
           climber().recordMiss();
         }
-        int oldWeightedSize = node.getPolicyWeight();
+        long oldWeightedSize = node.getPolicyWeight();
         node.setPolicyWeight(oldWeightedSize + weightDifference);
         if (node.inWindow()) {
           setWindowWeightedSize(windowWeightedSize() + weightDifference);
@@ -3485,7 +3485,7 @@ abstract class BoundedLocalCache<K, V> extends BLCHeader.DrainStatusRef
         boolean[] open = { true };
         return mappingFunction.apply(stream.onClose(() -> open[0] = false)
             .peek(ignored -> requireState(open[0], "stream has already been closed"))
-            .map(node -> nodeToCacheEntry(node, transformer, node.getPolicyWeight()))
+            .map(node -> nodeToCacheEntry(node, transformer, (int) node.getPolicyWeight()))
             .filter(Objects::nonNull));
       }
     } finally {
